Update test_view_limits_with_margins to account for other projections
This commit adds tests for margins on a symlog plot and a polar plot,
checking that the latter raises a NotImplementedError because it's
not implemented at the moment. It also uses approximate floating point
comparison for the linear and log plots to account for the fact that
the transformations are somewhat imprecise in the last bit.